+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.4.8 (2019/12/3)
 - Primary type supoort Long and BigInteger.
+- Correctly closed the Database Connection.
 
 ## 1.4.3 (2019/9/2)
 - Convert kotlin source of Query to java Source.